+++ dev_utils/management/commands/setupdevenv.py
@@ -9,6 +9,7 @@ from finance.models import Fee
 
 # THIS HAS TO BE REDONE
 
+
 class Command(BaseCommand):
     help = "Sets up environment for development (create Manager group, ...)"