+++ gig/forms.py
@@ -90,10 +90,22 @@ class StaffShowForm(forms.ModelForm):
 class GigLoadinChangeForm(BaseInlineFormSet):
     def clean(self):
         super(GigLoadinChangeForm, self).clean()
-        if 'cleaned_data' in self:
+        if self.is_valid():
             for form in self.cleaned_data:
-                if not form['DELETE']:
-                    self.instance.__loadin_depts__.append(form['department'])
+                if not form["DELETE"]:
+                    self.instance.__loadin_depts__.append(form["department"])
+
+
+class GigJobChangeForm(BaseInlineFormSet):
+    def clean(self):
+        super(GigJobChangeForm, self).clean()
+        for form in self.cleaned_data:
+            dept = form["department"]
+            valid_depts = self.instance.gig.__loadin_depts__
+            if dept not in valid_depts and not form["DELETE"]:
+                raise ValidationError(
+                    f"All jobs should match their parent systems or at least have a load in associated with their department. The current department choice for this job does not have a loadin."
+                )
 
 
 class GigSystemsChangeForm(BaseInlineFormSet):
@@ -103,16 +115,22 @@ class GigSystemsChangeForm(BaseInlineFormSet):
             system = form["system"]
             valid_depts = self.instance.__loadin_depts__
             if system.department not in valid_depts and not form["DELETE"]:
-                raise ValidationError(f"A loadin is required for each department that is working the show. Saving would create a situation where {system.get_department_display()} does not have a loadin.")
+                raise ValidationError(
+                    f"A loadin is required for each department that is working the show. Saving would create a situation where {system.get_department_display()} does not have a loadin."
+                )
 
 
 class sendStaffingEmailForm(forms.Form):
-    employees_working = forms.MultipleChoiceField(widget=forms.CheckboxSelectMultiple, label="", choices=["test", "test2"])
+    employees_working = forms.MultipleChoiceField(
+        widget=forms.CheckboxSelectMultiple, label="", choices=["test", "test2"]
+    )
 
     def __init__(self, *args, **kwargs):
         employees = kwargs.pop("employees")
         super().__init__(*args, **kwargs)
-        self.fields['employees_working'].choices = [(emp[3], f"{emp[0] if emp[0] else emp[1]} {emp[2]}") for emp in employees]
+        self.fields["employees_working"].choices = [
+            (emp[3], f"{emp[0] if emp[0] else emp[1]} {emp[2]}") for emp in employees
+        ]
         self.helper = FormHelper()
         self.helper.form_tag = False
         self.helper.disable_csrf = True