+++ unicorn/components/markTimesheetSigned.py
@@ -15,11 +15,11 @@ class MarktimesheetsignedView(UnicornView):
                 ts.signed = self.date
                 ts.save()
                 self.message = f"Timesheet {id} Signed with date {self.date}"
-                self.status = 'g'
+                self.status = "g"
             except TimeSheet.DoesNotExist:
                 if type(id) == int:
                     self.message = f"Timesheet {id} does not exist"
-                    self.status = 'b'
+                    self.status = "b"
                 else:
                     self.message = "This timesheet ID doesn't seem right...."
-                    self.status = 'b'
+                    self.status = "b"